Ported TestLocalDate to scalatest
