+++ jvm/src/test/scala/org/threeten/bp/TestLocalDateSerialization.scala
@@ -1,0 +1,69 @@
+/*
+ * Copyright (c) 2007-present, Stephen Colebourne & Michael Nascimento Santos
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *  * Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ *  * Neither the name of JSR-310 nor the names of its contributors
+ *    may be used to endorse or promote products derived from this software
+ *    without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+ * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+ * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+ * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.threeten.bp
+
+import java.io.{ByteArrayInputStream, ByteArrayOutputStream, ObjectInputStream, ObjectOutputStream}
+import java.lang.reflect.{Field, Modifier}
+
+import org.scalatest.{BeforeAndAfter, FunSuite}
+
+class TestLocalDateSerialization extends FunSuite with AssertionsHelper with BeforeAndAfter {
+  private val TEST_2007_07_15: LocalDate = LocalDate.of(2007, 7, 15)
+
+  test("serialization") {
+    val baos: ByteArrayOutputStream = new ByteArrayOutputStream
+    val oos: ObjectOutputStream = new ObjectOutputStream(baos)
+    oos.writeObject(TEST_2007_07_15)
+    oos.close()
+    val ois: ObjectInputStream = new ObjectInputStream(new ByteArrayInputStream(baos.toByteArray))
+    assertEquals(ois.readObject, TEST_2007_07_15)
+  }
+
+  test("immutable") {
+    val cls: Class[LocalDate] = classOf[LocalDate]
+    assertTrue(Modifier.isPublic(cls.getModifiers))
+    assertTrue(Modifier.isFinal(cls.getModifiers))
+    val fields: Array[Field] = cls.getDeclaredFields
+    for (field <- fields) {
+      if (!field.getName.contains("$")) {
+        if (Modifier.isStatic(field.getModifiers)) {
+          assertTrue(Modifier.isFinal(field.getModifiers))
+        }
+        else {
+          assertTrue(Modifier.isPrivate(field.getModifiers))
+          assertTrue(Modifier.isFinal(field.getModifiers))
+        }
+      }
+    }
+  }
+
+}
